Removed inode IDs from DirInode.ReadEntries.
The fuse documentation for readdir (http://goo.gl/fwHGtZ) says that
returning an entry for an inode ID doesn't affect its lookup count. This
makes it difficult to use our stateful approach for inode IDs, since
fuse doesn't tell us when to forget them. This implies to me that inode
IDs in responses to readdir are not important for use in applications.

I think this is correct -- when the user goes to open the file by path,
a LookUpInode request will later get hold of the real inode ID.