Removed `adduser $USER fuse` instructions for all but Debian/Ubuntu.
The `adduser` syntax is different on the other distributions (the
command doesn't even exist on sles-12). More importantly, installing
fuse doesn't seem to even create a group called `fuse`. When I Google
for context, the only instructions saying to add the user to the group
seem to all be related to Ubuntu and Debian, so this appears to only be
relevant there.

For #95.