Clarify that mode changes are silently ignored.
Currently the semantics are underspecified, and we actually return
ENOSYS in this case. But we need to silently ignore changes instead,
since otherwise the OS X Finder gives the user an error when they
attempt to copy in a file or directory (issue #125). I have observed
similar non-fatal errors printed on some Linux distributions.