Always request reads to the end of the object.
This doesn't actually copy the whole object, but fills the TCP buffers for
streaming. It's the equivalent of requesting a read of about 6 MB, and line the
client up for whatever data follows.

This is beneficial for streaming reads, especially those that don't begin at the
beginning of a file:
- Fewer requests to GCS reduces latency and increases throughput.
- Reduced financial costs of class A/B operations.

Signed-off-by: Bjorn Leffler <leffler@google.com>