Introduce BucketManager
Previously, gcsfuse sets up a gcs bucket and uses it to create the file
system. This is not feasible when the file system needs to support all
accessible buckets. Therefore, here comes the Bucket Manager, which
takes the params for the setup, and provides methods to set up new
buckets whenever they are needed.