Introduce inode.Name
Now, there is only 1 gcs bucket mounted at root, so the object name and
inode's local path match. However, when multiple buckets are mounted at
the subdirectories of the root, this is no longer true.

For instance, we may have bucket "foo" and "bar" mounted, and the files
are:
  - "./foo/x/y"
  - "./bar/x/y"
The gcs object name for both files would be "x/y". They use "x/y" to
talk to gcs APIs, but locally, it's named "foo/x/y" and "bar/x/y".

Since the gcs names and local names diverge, we introduce a name struct
to better manage such discrepency.