Load temp file on demand
Previously, when the temp file is created, it has to be loaded with the
source data in full extent. When `--local-file-cache` is specified, this
can be excessively slow and expensive, as all the files that are
detected will be downloaded, and operations will be blocked by the those
expensive downloads.

This commit eliminates the unnecessary downloads. The temp file will
only be filled with source data when it's being used.

TEST

- Run the unit tests in temp_file_tests.go.
- Run end to end tests with real gcs data.