Convert error to errno for FUSE
Before, an arbitrary error would be returned to FUSE from fs.go.
This is problemetic because FUSE only surfaces an syscall.EIO and
discards all the info, if the error is not an errno. For instance,
an googleapi.Error including a 403 (permission denied) would be
converted to "Input/output error" to the user side.

This change fixes it by adding some rules in errno() function to convert
the error to a proper errno.