Add the limit-mtime-mutation flag
When the mtime of a file is set, the FS updates the backing GCS
object's metadata directly. However, in some cases this leads to an
error due to insufficient permission for the object.update API. Thus,
this new flag is to limit this behavior. Once the flag used, the FS
ignores the mtime changes on unchanged files. It only sets mtime on
cached file content, which will be synced to GCS object at flush time.