Implement GetXattr and ListXattr to return ENODATA
Some clients, including Python scripts, would call `GetXattr` when open
files. Prior to this change, the Gcsfuse will print a lot of
```
ERROR: fuse.GetXattrOp error: function not implemented
```
which creates distractions for users to interpret the logs.

With this change, the GetXattr and ListXattr will always return ENODATA,
which gets ignored in the fuse connection.go. Therefore, they will not
be treated as an error. However, if the users try to SetXattr, it is
still considered an error and Gcsfuse prints a message for that.