Simply directory lookup by removing unnecessary ListObject requests
Previously, the inode lookup would call ListObject API to verify if a
child directory is implicit, which is expensive and unnecessary when the
directory already has a placeholder object. Such calls are removed here.

Benchmark shows that on n1-standard-96 using 8 threads reading 64 files,
this change will increase the bandwidth from 1.8GB/s to 2.4GB/s (+33%)
while reducing the frequency of ListObject calls from 339 to 66 (-80%).