Move the temp file destruction from syncer to inode
Each file inode keeps a temp file containing written data. The temp file
should be destroyed when the content is synced (uploaded), and other
conditions of the file inodes are met (such as local file cache
configs).

Before this change, the temp file destruction is triggered by the syncer
in the client library gcsx, regardless of the file inode conditions.

This change, however, moves the destruction out of the syncer to the file
inode, making it simplier to couple the destruction with the rest of the
inode instead.

Test:
```
go test ./...
```