Fix kokoro job failure for gzip integration tests
Kokoro VM is based on a specialized linux imaged by the Kokoro team for
running jobs on it. This has an old version of gcloud on it, and
thus the gcsfuse integration test jobs use `gcloud alpha` on it as
opposed to `gcloud` command.
To make matters worse, on the kokoro VM, `gcloud alpha` doesn't support
the `-Z` option as well as the options for updating object metadata.
Both these features are required for integration tests for GCS objects
with content-encoding gzip in GCSFuse. Hence, for both these features,
switching from `gcloud alpha` to equivalent `gsutil` commands.