Lower freq of syncs in periodic-perf-tests
This is for internal bug id 316284545 .
Currently in the periodic perf tests, we
fsync after each write (set through fsync=1)
in all write tests. This poses two problems.

1. It underrates our write bandwidth as most
of the runtime is spent in the sync calls and
little in the actual write operation.
2. This causes frequent sync which result in
frequent object generations on GCS, which is
throttled by GCS.

This change should address both these problems.