Switch statCache to internal.cache.Lru (#1514)
* statCache uses internal lru.Cache

This change switches internal.cache.metadata.statCache
to use gcsfuse-internal lru.Cache from
jacobsa/*lrucache.Cache .
This is a step towards unification of metadata cache with
the readonly cache interface and design. Currently
statCache takes in capacity (count) for initialization,
but is now well-placed to switch to size (in Mbytes) .

* minor cleanup

* Remove CheckInvariants() from StatCache

As new lru cache doesn't expose its
checkInvariants method, StatCache also
cannot implement a CheckInvariants()
method.

* rename invariantCache->testHelperCache

This change is there because there this helper class
no more does any CheckInvariants() call because it is not
supported anymore on statCache.