gargnitin/add-new-metadata-cache-config start
Add MetadataCache config

Add the support for following configuration fields in config-file.

```
metadata-cache:
	ttl-secs:
```

`ttl-secs` is an integer values denoting the ttl of stat-cache (aka
metadata cache in seconds). It is `-1` for no-ttl-based-expiry behavior,
`0` for no stat-cache, and `> 0` for normal ttl-based-expiry behavior.
A value of `ttl-secs < -1` will return error.
If ttl-secs is not defined, it will
fall back to the value of gcsfuse mount-flag `stat-cache-ttl`
(rounded up to the next multiple of second) .

If both ttl-secs and `stat-cache-ttl` are absent, then the default
value of 1 minute is used for ttl for stat-cache.

`stat-cache-ttl` will soon be deprecated.

This change also adds the corresponding unit tests etc.