Add metadata-cache:ttl-secs config (#1545)
Add MetadataCache config

Add the support for following configuration fields in config-file.

```
metadata-cache:
	ttl-secs:
```

`ttl-secs` is an integer values denoting the ttl of stat-cache (aka
metadata cache in seconds). It is `-1` for no-ttl-based-expiry behavior,
`0` for no stat-cache, and `> 0` for normal ttl-based-expiry behavior.
A value of `ttl-secs < -1` will return error.
If ttl-secs is not defined, it will
fall back to the value of gcsfuse mount-flag `stat-cache-ttl`
(rounded up to the next multiple of second) .

If both ttl-secs and `stat-cache-ttl` are absent, then the default
value of 1 minute is used for ttl for stat-cache.

`stat-cache-ttl` will soon be deprecated.

This change also adds the corresponding unit tests etc.

* Fix resource busy error in fs tests

* configure default values of *-cache-ttl

* set type-cache-ttl<= stat-cache-ttl

* Remove unwanted constant

* Merge DefaultTypeCacheTTL and DefaultStatCacheTTL

Merge DefaultTypeCacheTTL and DefaultStatCacheTTL into DefaultStatOrTypeCacheTTL

More minor changes e.g.
* change value of TtlInSecsUnset
* Updated some function headers

* Merge flags.(Stat|Type)CacheTTL into one

* fix a constant header

* update some headers

* Default ttl-secs=-1 to infinite-ttl

* address some review comments

* Updated semantics.md

* update more flags description

* update readcache mount_gcsfuse script

* update long-running test run_model script

* remove documentation for type cache capacity

* Move note about deprecation of old ttl flags down

* add a typo

* address a review comment