Add NestedSizeOfGcsObject (#1627)
* gargnitin/deep-size-of-utility start

* [wip] Add DeepSizeOf utility

DeepSizeOf utility recursively calculates the complete
size of a struct/built-in/pointer/slice/map etc.
This does not accurately represent size in memory,
because of missing accounting for memory alignments
and paddings, but is close.

* Revert "[wip] Add DeepSizeOf utility"

This reverts commit ebf4f5d273bc908776602f0ef1a97d84027e9ed6.

* Add NestedSizeOfGcsObject

This adds implementation of NestedSizeOfGcsObject by
recursively calculating and adding up the nested sizes
of all its members.
Currently it's defined in a hard-coded fashion (i.e. members are
used through names rather than fetching them by using reflect package.)
Generalizing this is left as an improvement idea for later.

* add first set of unit tests

* Make UnsafeSizeOf public

* add review comments from Prince

* addressed another comment

* remove nestedSizeOfString

* removed nestedSizeOfArrayOfStrings

* replace nestedSizeOfObjectAccessControl with contentSizeOfObjectAccessControl

* replace nestedSizeOfObjectAccessControlProjectTeam with contentSizeOfObjectAccessControlProjectTeam

* clean-up unnecessary vars from sizeof_test.go

* fix a header

* rename tests