Add type-cache composite tests (#1634)
* gargnitin/type-cache-composite-tests/v1 start

* Add debug logs to cache.LRU

To be removed later

* Revert "Add debug logs to cache.LRU"

This reverts commit 9d3354eab718ef5527ea7cf73a6f7ce5c6b04d7a.

* Add debug logs to type-cache

To be removed later...

* Revert "Add debug logs to type-cache"

This reverts commit adc3f45d9ed6900e8824eb4e37d00100055e6ef2.

* Consume t.servercfg.mountConfig in fs_test.go

This is needed to have the tests derived from
fsTest to use the InodeCacheTTL, DirInodeTTL, logging
related config from the t.servercfg.mountConfig , needed for
tests which need to test configuration through config-file.

* Add type-cache composite tests

This adds some composite tests for
type-cache along the same lines as the other
tests in internal/fs/*_test.go .
This is not an exhaustive set, but a first
batch, and more might be added later.

* Update type-cache composite tests for max-size-mb

Updating for type-cache-max-size-mb

* Remove log-file-config from internal/fs/type_cache_test.go

* Remove log-file-init from fs_test.go

* fix test broken by rebase

* Reduce runtime of type-cache composite tests

Also move mountConfig -> serverCfg to type_cache_test.go

* be stattable -> pass stat call

* add explanation of how os.Stat works

* Add parellization in TypeCacheTestWithMaxSize1MB.TestSizeBasedEviction

* Added tests for validating stat behavior

* De-dup code by creating utility functions

* fix comments and some names

* remove unintended binary file

* fix test broken by rebase