Downgrade aws-sdk-go to 1.44.217
Downgrading aws-sdk-go from 1.50.x back to 1.44.217.
Because of the rebase conflicts and because of the
intermediate upgrade of aws-sdk-go to 1.50.x, the version
got stuck at 1.50.x .
In the current latest master and read_cache_release
branches both, the version is at 1.44.217, so lowering
it here too to 1.44.217 for consistency.