Call ls -R from within mount
This call runs `ls -R` on mount-directory,
right after returning successful mount to user.
This initiates the `ls -R` in a separate
goroutine. On completion of this goroutine,
if there is an error, then it is logged
in gcsfuse debug log-file.
The `ls -R` is simulated with filepath.WalkDir
which recursively visits each sub-directory
and takes count of all files/directories.