Metadata-prefetch (aka Recursive listing or ls -R during mount) (#1930)
* Call ls -R from within mount

This call runs `ls -R` on mount-directory,
right after linux mount, and blocks
returning after mount completion to the user
until the `ls -R` completes.
The `ls -R` is simulated with filepath.WalkDir
which recursively visits each sub-directory
and takes count of all files/directories.

* Temp - [WIP] concurrent listings of sibling directories

* Revert "Temp - [WIP] concurrent listings of sibling directories"

This reverts commit c34e77f09dfc995769ff5a004a127c2c21d0c76e.

* Add mount-config metadata-prefetch-mode

Added mountconfig-flag "metadata-prefetch-mode'
for config-file. The type of value passed is a string.
The supported values are: (a) disabled, (b) synchronous, (c)
asynchronous.
If not set, the value `disabled` is taken by default.

If value `synchronous` is set, then mount is blocked
until the metadata prefetch is complete. In this mode, if
the metadata-prefetch fails, the mount itself fails.

If value is set to `asynchronous`, then mount is not
blocked, but metadata prefetch is
initiated in a go-routine and will complete sometime
later. In this case, if the metadata-prefetch fails,
then an error is logged in the GCSFuse log-file
set using log-file in commandline flags, or file
in `log-config: file` in config-file.

Metadata-prefetch is equivalent to running `ls -R`
on the mount-directory to prime/prefill/prefetch the
metadata caches (both type and stat).

* minor improvements

* Migrate main_test.go to stretchr/testify

* metadata-prefetch-mode configs in listing e2e tests

* fix lint errors

* temp-fix for integration-test failure

* fix some self-comments

* add unit tests for populateList

* handle dynamic-mount scenario

* beautify test code

* increase code-coverage

* add code for coverage

* fix typo in prev commit

* remove type metadataPrefetchModeValue

* address a review comment

* Add commandline flag metadata-prefetch-mode

And switches main.go to call metadata-prefetch based on this flag from the mount-config flag of the same name.
Also addresses some minor code comments.

* Put back unintentionally ignored error

* Migrate flags_test to stretchr/testify

* remove config-flag metadata-prefetch-mode

* fix for persistent mounting

* address couple of review comments

* addressed some more comments

* add linter-error abt missing error-check

* Rename metadata-prefetch-mode

.. to metadata-prefetch-on-mount

* Change [a]synchronous to [a]sync for md-prefetch

Change "synchronous" to "sync", and
 "asynchronous" to "async" for values for
 metadata-prefetch-on-mount everywhere.

* address review comments

* MetadataPrefetchMode -> MetadataPrefetchOnMount

Changes in all internal state variables/ constants etc. needed to fix the mount-flags log.

* Optimize code

This creates a mini-race condition between the
debug log for successful and that of starting
the recursive listing of mountPoint when
metadata-prefetch-on-mount has been set to "async",
but it is immaterial to the end user.

* handle _ as bucket for dynamic-mount

* optimize code

* improve error logs

* Remove callList from daemon-parent

Removed redundant call to callRecursiveList
from the parent process of the GCSFuse daemon.

* Code optimization

Reuse isDynamicMount utility in mount.go