Don't compute coverage while running tests (#1969)
The coverage computation might be resulting in the transient failures that we have been seeing in CacheRaceDetector tests. Refer: https://youtrack.jetbrains.com/issue/GO-2037/Running-Tests-with-coverage-and-Go-tool-arguments-race-gives-erroneous-race-conditions.
Since we already compute coverage as part of codecov.yml, and we are anyway not using the code-coverage file generated as part of the tests, removing it should not have any side-effects.