Support higher precedence for config-params
In order to be backward compatible, we need to handle cases where the
config-file params have higher precedence than their CLI counterparts.
For those cases, we introduce another viper called the config-viper that
sources values from the config-file alone (+ the default value). We can
use the config-viper.IsSet(<key>) to figure out whether the config file
has the key explicitly specified and if so, use the object that's been
unmarshalled from the config-viper.
The other viper sources values from both CLI as well as config-file. So,
that can be used in cases where CLI flags should have a higher
precedence or as a fallback for cases where Config file should have a
higher precedence but the value is not set.
We plan to move to a state where CLI has higher precedence compared to
config file in all cases. So, config-viper should go away in the next
major release.