Handle version and help flags
urfave/cli uses the golang flags library which treats the "-flagName" and "--flagName" as equivalent, however Cobra, which uses the pFlags (POSIX flag standard), treats single-hyphen-prefixed flags as shorthands while doubly-hyphen-prefixed as flags. In order to be backward compatible, we prefix the flags that have been specified with a single hyphen with another hyphen. This however causes problems for certain shorthand flags such as "-v" and "-h" which have special meaning to Cobra. Prefixing them with another hyphen causes errors since to Cobra, Cobra only knows about "v" and "h" shorthands not flags.
As part of this change we address the following:
* Don't modify "-h" to "--h"
* Convert "--h" and "--v" to "-h" and "-v" respectively.