Make Logger an exported param
Currently, all the logger calls create a string out of the specified
format string and values. However, this means that one cannot avail the
benefit of structured logging where one can pass key - value pairs to
the log which structures them as fields and values instead of encoding
them all as a plain string. This structure in the logs allows for better
data analysis without resorting to complex string processing.
Exposing the sLog handle allows users to use the structured logging
APIs.