Enable kernel symlink caching
This would provide significant savings when symlink'ed path is
frequently accessed.

Since symlink target is not mutable, this is always safe regardless TTL
settings. (When the symlink is removed and re-created, its generatiton
would change so a different inode is created). However, as a precaution,
gate this under a flag for now.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>