escape comma in gcsfuse mount options
This is needed as helm takes comma(,)
in the helm argument values as a separator
for values. This is a problem for gcsfuse-mount-options
as they may contain commas and are passed to helm as
a single argument. So it is necessary to
escape them while passing the value to helm.