Remove fs-error label (#2382)
fs-error has a high cardinality and will not be supported by GCM.
So, either we can drop it in gcsfuse or gcs-fuse-csi-driver will have to make
changes to drop the label. fs-error is likely not needed since we
already are classifying the most frequent errors into a new label called
fs-err-category and the other errors are being classified into
appropriate buckets. Moreover, we'd continue to have logs if we need to
get more details about an (infrequent) error. So, we can remove fs-error
label.