Fix flakiness in tracing test.
The "Tracer" objects are associated with a TracerProvider. So, if the
TracerProvider changes, the pre-existing "Tracer" objects continue to
log to the older TracerProviders. In the monitoring.go, the code was
only initializing the Tracer once as a var at the top of the file, which
has the following implications:
* If Tracer is initialized before the global TracerProvider is set,
  it'll be a no-op and the traces will get missed.
* If the same test is run multiple times and since each time a new
  TracerProvider is created, the older Tracer becomes a no-op and the
  subsequent tests will fail.

To fix this, create the Tracer from the latest TracerProvider everytime.