Add shadow reviewer when PR's review is requested (#2505)
ready_for_review triggers the GitHub action only when a PR is moved from draft PR to "Ready to Review" stage.
So, shadow-reviewers don't get added when a PR is directly created without going through
the draft stage. Use review_requested to add shadow reviewer when a review is requested.