[testing-on-gke] Fixes and improvements in the test setup (#2329)
Break-down of changes:
* This removes dependence on user having to pass absolute paths for arguments in the run-script.
* Passes gcsfuse-mount-options, machine-type from run-script down to the python run script for better programmability 
   rather than manipulating pod configs using SED.
* Uses kubectl namespace properly in the run-script.
* Makes output file/dir configurable.
* Added overall timeout on pod runs.

Dirty notes of the constituent changes:

* dont switch branch on preexisting code

Use the give branch i.e. gcsfuse_branch
env var only when cloning code afresh.
If user passed gcsfuse_src_dir, don't
checkout gcsfuse_branch on it.

* Pass gcsfuse_mount_options, machine_type down

... from run-gke-tests.sh to run_tests.py.

* Remove pod yaml update by sed

* escape comma in gcsfuse mount options

This is needed as helm takes comma(,)
in the helm argument values as a separator
for values. This is a problem for gcsfuse-mount-options
as they may contain commas and are passed to helm as
a single argument. So it is necessary to
escape them while passing the value to helm.

* take workload_config as argument to run-script

* Add output_dir argument in run-script

* add instruction to connect to any pod

* Resolve all paths passed to absolute paths

* Improvements

Remove some commented out, unnecessary lines.
Move CSI code fetch dependency to csi-custom-driver-build path.
Remove some unnecessary function calls.

* fix logs about csi driver enabling/disabling

* use namespace properly in run-script

* Address review comments

* Fix some docstrings
* Add some missing input checks

* Minor fixes

1. Remove check of pre-existence of implicit workload_config/output_dir.
2. Minor redundant code removal

* Address review comments

* Add pod timeout

Add an overall pod-timeout to
kill the run if the pods fail or
get stuck somehow.

* increase default-timeout to 1wk