Propagate the modified context (#2476)
OTel tracing leverages context to propagate information about the
ongoing trace. While the invokedWrapper method creates a new context
as part of creating a new trace, it doesn't propagate that context to
the subsequent method calls; the old context gets propagated. This
causes the subsequent spans to create a new trace instead of joining the
ongoing one. Fix that by propagating the context containing span
information.