[gke-automated-test-tool] fix bug in csi driver build (#2819)
This fixes a bug unintentionally introduced in previous commit . In that commit, the bash function uuid was printing out some more information other than the uuid value itself because of the installation step. In bash, the printed log becomes the output of a function, hence the installation inside this function was changing the output of the uuid function breaking the caller. So, I have moved the installation out to wherever we're invoking this function.