Enable dynamic creation of storage clients based on bucket type and specified protocol (#2922)
Enable dynamic creation of storage clients based on bucket type and specified protocol

Reason for not using mutex in getClient:
Multiple clients can be created in dynamic mounting case only. But in this case, the root dir will actually be a base_dir & we take a lock before calling LookUpChild for base_dir. So, this will ensure that call to BucketHandle (which means calls to getClient) will be serialized.

Reason for not adding a new config EnableBidiConfig in StorageClientConfig (& instead passing a separate parameter for bidi config while creating grpc client): 
This config is created from the user provided config at the start (& should reflect that) and I would not like to keep a transient field to it (i.e. would not like to keep a field which has to be set every time before grpc client creation happens and does not serve any other purpose apart from reducing an extra parameter to grpc client creation call).

Testing: Manually tested mounting/listing/reading on flat, hns and zb bucket with this change using dynamic mounting (with both grpc as well as http)