Change to properly parse the error (#2953)
gRPC returns status messages on error (https://google.aip.dev/193#error_model). So, it should be used to handle gRPC errors.

apiError is a wrapper over status & googleApi.Error. So, ideally, we should just be using apiError.Error and then check whether the underlying error was http or gRPC and then proceed (i.e. try to convert error to apiError.Err and if it succeeds, then use appropriately). But the http error that we got returned success when converted to both status as well as googleApi.Error (although with unknown code in converted status).

Hence, we have to check for both scenarios (http & grpc). So, there is no use of using apiError. Hence, started using status directly for gRPC.