Fix metrics performance
The gcs/read_bytes_count metric is single-handedly responsible for perf
degradation due when metrics are enabled. With this change, we are now
using an atomic Int64 which efficiently computes the metric. The
Int64Counter that uses OTel for metric computation has now been replaced
with Int64ObservableCounter that uses the aforementioned atomic Int64 to
report the same metric.