[testing-on-gke] Implement flag custom_csi_driver
custom_csi_driver is a new env-var based
option for fio/dlio GKE test tool,
which allows user to pass a pre-built
custom csi driver, rather than having
to build a new one each time.
It works in conjunction with use_custom_csi_driver.
If a non-empty custom_csi_driver is passed,
then the value of use_custom_csi_driver is overridden
and the passed custom-csi-driver is passed. If
is not passed, or passed as "", then a new custom
csi driver is built if use_custom_csi_driver is
passed as true (legacy behaviour).

This commit doesn't really implement the
functionality of custom_csi_driver, but only
introduces the flag. The implementation will be in
later commits.