Unify the code-flow with and without metrics being enabled. (#3288)
Currently, enabling metrics will conditionally enable a decorator that activates the code flows to record metrics. However, this leads to test-gaps since we currently don't enable metrics in our tests and that decorator doesn't come into picture while running tests. So, we unify the flow with/without metrics enabled so that the monitoring decorator will always be set even in cases when no metrics are to be computed.