Fix broken support for requester-pays buckets (#3256)
Pass billing-project in every GCS call, except LROs (Long Running Operations) from storage control-client (through "x-goog-user-project") to fix mounting and operations of requester-pays buckets . This functionality was broken since v2.9.0 .

This fixes the mount of requester-pays buckets, as well as the following operations which go through storage-control-client.
- GetFolder,
- CreateFolder,
- RenameFolder # might fail even after this change until the bug fix for GCS bug (b/414790551) completely rolls out.
- DeleteFolder

In terms of code, this also required
- Add a new implementation of the interface StorageControlClient for the case when a billing-project has been specified by the gcsfuse user.
Note: After this PR merged, E2e tests need to be added for requester-pays buckets. That task is captured in b/415244758

Link to the issue in case of a bug fix.
b/406090515, b/405913611

* pass x-goog-user-project to GetStorageLayout call

* dont pass empty billingProject to GetStorageLayout

* Revert "dont pass empty billingProject to GetStorageLayout"

This reverts commit 62ddbce573f27e152ca5888056012367bc191bba.

* Revert "pass x-goog-user-project to GetStorageLayout call"

This reverts commit 74fc8917f1e139a9a60b386f76dfbd7a7044cec3.

* [unstable] Use biling-project in control-client: part 1

* Use biling-project in control-client: part 2

* fix header

* fix lint errors

* empty commit

* reverted all prev commits at once

* Add cleaner, shorter code

* add unit test

* empty commit

* Fix failure in emulator_tests

* dont pass billing-project in RenameFolder calll LRO