Fix dynamic-mounting test in "operations" package for ZB (#3187)
This addresses a TODO left behind from couple of comments (comment1 and comment2) in PR #3172 . The issue was that the bucket created for the dynamic bucket test was creating a non-ZB earlier which returned storage-class as "STANDARD" which mismatched with our test which was checking for it to be "RAPID". I replaced this code to create a ZB instead of non-ZB in case of ZB test runs.

Link to the issue in case of a bug fix.
b/409734728

Testing detail

* create zonal bucket for dynamic mounting

* fix zonal bucket creation

* terminate test if file open fails

* comment out added sleeps between creates and opens

* change bucket name for ZB

* add an important TODO for future

* disallow false-pass for ZB in dynamic mount test

* Revert "change bucket name for ZB"

This reverts commit a8aded9809b4ea1f8a179609e374a001778f1b5a.

* remove some commented out debug code

* Revert "add an important TODO for future"

This reverts commit 33f9d2b949099d9af20524da28e46d67db4313ac.

* removal commented out line