Exclude files from cache based on name (#2043)
* Exclude files from cache based on name

The current logic to assume that files are meant for small random reads
if the first read to them has a non-zero offset assumes the absence of
file headers at the beginning of the file, and/or the absence of magic
number checks.

The new option to specify such files by regular expression pattern will
allow the users to mark such files as random access only, even when
their headers are fetched first.

* Review comments

* Sync new flag documentation with config yaml

* fix tests