feat(parallel random reads): Changes to enable parallel random read handling in gcs reader (#3619)
Changes to enable parallel random reads in the GCS reader. This is achieved by refactoring the read logic in gcs_reader.go, introducing a mutex to manage concurrent access for range reader while allowing reads to go in parallel for MRD. The changes also include updates to the test suite to cover the new logic and parallel read scenarios.

Since, we're removed the high level fh lock for read operation (ReadFile in fs.go), we've added extra synchronization logic in file handle operations.