fix: Disable e2 test for renaming unfinalized objects (#3707)
* Disable test for renaming unfinalized objects

The support for renaming unfinalized objects is there
in some cloud regions, but not in others.
So, this test (TestUnfinalizedObjectCanBeRenamedIfCreatedFromDifferentMount)
fails in some regions and passes in others.
To be consistent across all runs in all regions,
I am disabling this test. Will re-enable it when
rename for unfinalized objects is there consistently
everywhere.

* add todo comment for reverting