Add DNS caching support in GCSFuse
GCSFuse doesn't always read the entire data off of the Readers that it creates. Due to this,
the HTTP/1 connection doesn't get reused. Every new connection requires a DNS call which in
certain cases such as host-network could put a lot of load on the metadata-server. Fix that
by implementing a DNS cache within GCSFuse.