feat: Add mount-id as a resource attribute for metrics
This change adds the mount-id as a resource attribute to the OpenTelemetry metrics exporter. This allows for the differentiation of metrics from multiple gcsfuse mounts. Additionally, the PID is removed from the metrics attributes.

The mount-id is passed from the Mount function to the SetupOTelMetricExporters function, which then adds it as a service.instance.id attribute to the OpenTelemetry resource. The PID is removed from the WithFilteredResourceAttributes option in the setupCloudMonitoring function.