fix: Propagate OTel tracing context in fs-ops (#3940)
### Description
The OTel trace instrumentation setup in internal/fs/wrappers/tracing.go is not propagated when ignoreInterrupts is set default to true in internal/fs/fs.go file system interface methods. Because of this the trace context from WithTracing wrapper is lost and underlying libraries like go storage grpc instrumentation is not tracked under the same trace. This PR ensures the span context from parent is still passed down when tracing is enabled.

### Link to the issue in case of a bug fix.
b/454805201

### Testing details
1. Manual - DONE

The following test plan was followed to test tracing is properly working as expected.
Enable experimental monitoring/tracing config with sampling ratio 1 to track all traces and export them to gcloud.
Using the mount instance id, filter the traces and check the trace for any of the file system methods like WriteFile or LookupInode. 
Compare these trace results with and without fix and it can be seen that the trace is as expected with the fix.

Screenshots below:

Trace with fix:

<img width="1621" height="344" alt="trace_with_fix" src="https://github.com/user-attachments/assets/82446148-ce70-4309-b913-263b2f8d11a9" />

Trace without fix:

<img width="1628" height="407" alt="trace_without_fix" src="https://github.com/user-attachments/assets/f2cd3952-9b88-4d24-96bc-5c6e2790f32d" />


2. Unit tests - Added
Added unittests for all the fs ops method in fs/fs.go using fake bucker manager to check spans generated.

3. Integration tests - NA

### Any backward incompatible change? If so, please explain.
NA