fix(monitor): disable Cloud Monitoring exporter on permission denied error (#4191)
When the service account that runs GCSFuse with cloud-metrics-exporter enabled but lacks the monitoring/metricWriter permission, then currently, the exporter continuously attempts to export and fails causing an error log for every instance of failed export. This PR fixes it by disabling the exporter when a permission-denied error is detected.